+++ packages/styles/src/index.ts
@@ -80,7 +80,6 @@ import '@bkui-vue/search-select/search-select.less';
 import '@bkui-vue/config-provider/config-provider.less';
 import '@bkui-vue/pop-confirm/pop-confirm.less';
 import '@bkui-vue/image/image.less';
-import '@bkui-vue/image/image-viewer.less';
 import '@bkui-vue/overflow-title/overflow-title.less';
 import '@bkui-vue/scrollbar/css/scrollbar.less';
 import '@bkui-vue/directives/tooltips.less';