+++ site/views/image/demo/error.vue
@@ -1,0 +1,7 @@
+<template>
+  <bk-image
+    :height="120"
+    :width="200"
+    src="/not-exist.jpg"
+  />
+</template>